requirements.txt
Django==5.0.2
djangorestframework==3.14.0


python-dotenv==1.0.1

pandas==2.2.0
numpy==1.26.3
scikit-learn==1.5.2
joblib==1.3.2

ortools>=9.8.3296

python-multipart==0.0.6
reportlab>=4.0.0
Pillow>=10.0.0

# Charts/UI
jinja2==3.1.3

# Optional but helpful
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.6.0